Extract timeline item and mask editor helper logic
